requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
html5lib
tqdm==4.66.1
openai==1.0.0
regex
python-dotenv==1.0.0
